Updated database-test.js
Changed database-test.js so you can now input a URL you want to insert into the URLCollection in the database.
This is done using the insert function.
Also found a bug that occurs when you call the insert function and the read function in any sequence. For some reason connection to the database does not close. Still working on solution.