Added index.js and saveResults.js
Added two new files to the Webacraper folder. I'm planning on using these to separate the Web Scraper program into parts. The webscraper.js will scrape the data. The saveResults.js will send the data to the database. Index.js will be used to start the web scraper.
I'm not sure if I will follow through with this, it's just an idea for now.