Fixed Web Scraper bug
Fixed a problem I encountered with the web scraper program. It appears that the CSS selectors I used previously were no longer correct. The problem was fixed when I went back to the Twitch website and manually copied the CSS selectors again.
My theory is that when Twitch updates their website the selectors may change. We may have to keep this in mind as we test the web scraper.