Added database read and insert test files
Added two new files to the web scraper.
The database-readURL.js file is used to print out all of the URLs stored in the URLCollection of the database.
The database-insertURL.js file is used to insert a URL into that same collection.
These two files are used to manually test the web scraper and make sure the connection to MongoDB works.