Added two testing files to the backend
Added a testing.js file that I am using to implement the organizing feature of the web scraper.
Added a file named database-renameCollection.js that can be used to rename collection in the database. This file was added because some collections in the database needed to be renamed to match the category name that can be seen when inspecting the Twitch website.