Added organizing functionality to webscraper.js
Added organizing functionality to the web scraper. The web scraper will now check what category a stream has. If it is a category we have approved, then the web scraper will send the stream data to the database. Else, the category is unapproved so the web scraper moves onto the next URL link.

The web scraper will also organize the stream data it gathers into the appropriate collection in the database.

I will now move on to combining this functionality with the version of the web scraper that listens to the database with change streams.