Combined organizing and listening
Combined the stream organizing and database listening functionality into a single web scraper. This web scraper is the one located in the webscrapertest.js file.
This web scraper will listen to the URLCollection of the database, and when a new URL is inserted it will start the web scraping process. The web scraper is able to tell if stream data should be sent to the database, and what collection to put it in.
I plan to move this web scraper into the webscraper.js file after further testing.