Added live viewer count verification
Implemented a web scraper that checks if a stream if live, the stream category, and stream live viewer count. This web scraper verifies streams before they get sent to the main web scraper.
If a stream is not live then it will not be sent. If the live viewer count of a stream if greater than 100 it will not be sent.
If a stream has a category that is not listed on the website it will not be sent.