Delete functionality for main web scraper
Added delete functionality to main web scraper. Once it gets data from a stream it will remove the stream link from the VerifiedURLCollection.
The main web scraper also has backup checks to stream category and if a stream is live.
If the stream category is not listed on the website the web scraper will remove that stream URL from the VerifiedURLCollection. It will do the same for streams that are not live.